+++ hardhat-config/external.config.js
@@ -4,6 +4,12 @@ module.exports = {
       {
         artifacts: 'node_modules/@animoca/ethereum-contracts/artifacts',
       },
+      {
+        artifacts: ['node_modules/opencampus-xyz/EDU-smart-contract/artifacts'],
+      },
+      {
+        artifacts: 'node_modules/@animoca/anichess-ethereum-contracts-2.2.3/artifacts',
+      },
     ],
   },
 };