+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 
 # Changelog
 
+## 1.2.1
+Updated OCPointMerkleClaim smart contract and related test cases based on audit report issued on 2024/10/28
+- AB-SC-SC-01: Incorrect variable OCPoint casing
+- AB-SC-SC-02: Comments do not correctly describe the contract’s behavior
+
+## 1.2.0
+- Added OCPointMerkleClaim smart contract and related test cases.
+- Add "@animoca/anichess-ethereum-contracts-2.2.3" as devDependencies
+
+## 1.1.0
+
+Added escrow reward related smart contracts and related test cases.
+- PublisherNFTEscrow
+- GenesisTokenEscrow
+- EDuCoinMerkleClaim
+
 ## 1.0.0
 
 Initial release.